split spawnBinaryProcess() into spawn, no changes
